release1:take a rest and go to play rust.

+++ testIo/testByteArrayReader/ByteArrayReaderBigEndian.cpp
@@ -102,7 +102,7 @@ void testByteArrayBigEndian() {
         ByteArrayReader reader = createByteArrayReader(testData,st(IO)::Endianness::Big);
         
         auto v1 = (array2[0]<<8 | array2[1]);
-        int v2_chk = reader->read<int>();
+        int v2_chk = reader->read<short int>();
         if(v1 != v2_chk) {
           TEST_FAIL("[TestByteArrayReader BigEndian case9],v1 is %d,v2_chk is %d",v1,v2_chk);
         }
@@ -115,7 +115,7 @@ void testByteArrayBigEndian() {
         ByteArrayReader reader = createByteArrayReader(testData,st(IO)::Endianness::Big);
         
         auto v1 = array2[0];
-        int v2_chk = reader->read<int>();
+        int v2_chk = reader->read<byte>();
         if(v1 != v2_chk) {
           TEST_FAIL("[TestByteArrayReader BigEndian case10],v1 is %d,v2_chk is %d",v1,v2_chk);
         }
@@ -128,7 +128,7 @@ void testByteArrayBigEndian() {
         ByteArrayReader reader = createByteArrayReader(testData,st(IO)::Endianness::Big);
         
         auto v1 = (array2[0]<<8|array2[1]);
-        long v2_chk = reader->read<long>();
+        long v2_chk = reader->read<short int>();
         if(v1 != v2_chk) {
           TEST_FAIL("[TestByteArrayReader BigEndian case11],v1 is %d,v2_chk is %d",v1,v2_chk);
         }
@@ -140,7 +140,7 @@ void testByteArrayBigEndian() {
         ByteArray testData = createByteArray(array2,2);
         ByteArrayReader reader = createByteArrayReader(testData,st(IO)::Endianness::Big);
         
-        auto v1 = reader->read<int>();
+        auto v1 = reader->read<short int>();
         int v1_chk = (array2[0]<<8|array2[1]);
         if(v1 != v1_chk) {
           TEST_FAIL("[TestByteArrayReader BigEndian case12],v1 is %d,v2_chk is %d",v1,v1_chk);
